+++ src/router/index.ts
@@ -4,7 +4,7 @@ import StudentHomeView from "../views/student/Home.vue";
 import LecturerHomeView from "../views/lecturer/Home.vue";
 import AdminHomeView from "../views/admin/Home.vue"
 import LecturerEditCreateCourseView from '../views/lecturer/EditCreateCourseForm.vue';
-import AdminCreateAccountView from '../views/admin/CreateAccountForm.vue';
+import AdminEditCreateAccountView from '../views/admin/EditCreateAccountForm.vue';
 import Redirect from "../views/common/Redirect.vue"
 
 
@@ -49,12 +49,12 @@ const router = createRouter({
 		{
 			path: "/createAccount",
 			props: {editMode:false},
-			component: AdminCreateAccountView
+			component: AdminEditCreateAccountView
 		},
 		{
 			path: "/editAccount/:username",
 			props: {editMode:true},
-			component: AdminCreateAccountView
+			component: AdminEditCreateAccountView
 		},
 		{
 			path: "/redirect",
